feat: pass human-readable experiment names to RL module
- Update ExperimentRegistrationService to accept optional experiment name
- Update ai_sorting.module to pass view:display format experiment names
- This ensures experiments display as readable names in RL reports

Fixes issue where experiments showed SHA1 hashes instead of readable names
in reports. Now experiments show as "view_name:display_name" format making
them much more user-friendly.
