--- ai_sorting.permissions.yml
@@ -1,3 +1,0 @@
-administer ai sorting:
-  title: 'Administer AI Sorting'
-  description: 'Configure AI sorting parameters in Views'
