+++ ai_sorting.services.yml
@@ -1,17 +1,4 @@
 services:
-  ai_sorting.experiment_resolver:
-    class: Drupal\ai_sorting\Service\ExperimentResolver
-
-  ai_sorting.total_trials_service:
-    class: Drupal\ai_sorting\Service\TotalTrialsService
-    arguments: ['@rl.experiment_manager']
-
-  ai_sorting.experiment_decorator:
-    class: Drupal\ai_sorting\Decorator\AiSortingExperimentDecorator
-    arguments: ['@entity_type.manager']
-    tags:
-      - { name: rl_experiment_decorator }
-
   ai_sorting.experiment_registration:
     class: Drupal\ai_sorting\Service\ExperimentRegistrationService
     arguments: ['@rl.experiment_registry']
