--- docs/ai_sorting_project_desc.html
@@ -1,98 +1,0 @@
-<p>This module is part of the <a href="https://www.drupal.org/project/ai">AI module</a> ecosystem and included in <a href="https://www.drupal.org/project/dxpr_cms">DXPR CMS</a>.</p>
-
-<blockquote>Transform your content marketing with AI Sorting - the Drupal module that automatically learns which content engages your audience and promotes it for maximum impact. Stop guessing what works and let machine learning optimize your content visibility.</blockquote>
-
-<h3>You need AI Sorting if</h3>
-<ul>
-  <li>You want your best-performing content to automatically rise to the top</li>
-  <li>You need machine learning to optimize content visibility without manual work</li>
-  <li>You manage blogs, news feeds, or product catalogs with varying engagement</li>
-  <li>You want real-time content optimization based on user interactions</li>
-  <li>You prefer data-driven content ranking over manual curation</li>
-</ul>
-
-<h2>About AI Sorting</h2>
-
-<h3>Boost Your Content Performance</h3>
-
-<ul>
-<li><strong>Automatic Optimization</strong> - Content that gets clicks rises to the top</li>
-<li><strong>Smart Discovery</strong> - New content gets fair exposure while proven winners shine</li>
-<li><strong>Real-time Learning</strong> - Adapts continuously as user preferences change</li>
-<li><strong>Zero Maintenance</strong> - Set it once and watch it work automatically</li>
-</ul>
-
-<h3>Perfect for Content Marketers</h3>
-
-<ul>
-<li><strong>Blog Posts</strong> - Surface your most engaging articles first</li>
-<li><strong>Product Pages</strong> - Highlight items that convert best</li>
-<li><strong>News Feeds</strong> - Breaking stories get automatic priority</li>
-<li><strong>Resource Centers</strong> - Most valuable downloads rise to the top</li>
-<li><strong>Case Studies</strong> - Showcase your most compelling success stories</li>
-</ul>
-
-<h3>The Marketing Advantage</h3>
-
-<p>Unlike static content lists, AI Sorting delivers:</p>
-
-<ul>
-<li><strong>Higher Click-Through Rates</strong> - Engaging content gets prime real estate</li>
-<li><strong>Better User Experience</strong> - Visitors find relevant content faster</li>
-<li><strong>Reduced Bounce Rates</strong> - Keep audiences engaged with smart recommendations</li>
-<li><strong>Content ROI Insights</strong> - See which content truly performs</li>
-<li><strong>Evergreen Optimization</strong> - Content performance improves over time</li>
-</ul>
-
-<h3>How It Works</h3>
-
-<ol>
-<li><strong>Track Engagement</strong> - Monitors when content is viewed and clicked</li>
-<li><strong>Learn Patterns</strong> - Machine learning identifies high-performing content</li>
-<li><strong>Optimize Order</strong> - Best content automatically moves to prominent positions</li>
-<li><strong>Continuous Improvement</strong> - Performance gets better with every visitor</li>
-</ol>
-
-<h3>Marketing-Focused Benefits</h3>
-
-<ul>
-<li><strong>Content Performance Analytics</strong> - Built-in reporting shows what resonates</li>
-<li><strong>Audience Behavior Insights</strong> - Understand what drives engagement</li>
-<li><strong>Campaign Optimization</strong> - Promoted content gets measured and optimized</li>
-<li><strong>Seasonal Adaptation</strong> - Content relevance shifts automatically with trends</li>
-</ul>
-
-<h3>Simple Setup</h3>
-
-<ol>
-<li><strong>Install</strong> - One-click installation with the RL module</li>
-<li><strong>Configure</strong> - Add AI Sorting to any content listing in minutes</li>
-<li><strong>Launch</strong> - Start collecting engagement data immediately</li>
-<li><strong>Optimize</strong> - Watch your content performance improve automatically</li>
-</ol>
-
-<h3>Enterprise-Ready</h3>
-
-<ul>
-<li><strong>High-Traffic Optimized</strong> - Handles enterprise-scale content volumes</li>
-<li><strong>Performance Balanced</strong> - Smart caching for speed without losing accuracy</li>
-<li><strong>Mobile Responsive</strong> - Works across all devices and platforms</li>
-</ul>
-
-<div class="note-tip">
-  <h4>Prefer a turnkey demo site?</h4>
-  <p>Spin up <strong>DXPR CMS</strong>—Drupal pre-configured with DXPR Builder, DXPR Theme, AI Sorting module, and security best practices.</p>
-  <p><a href="https://www.drupal.org/project/dxpr_cms" title="DXPR CMS platform">Get DXPR CMS »</a></p>
-</div>
-
-<h3>Dependencies</h3>
-
-<ul>
-<li><a href="https://www.drupal.org/project/rl">RL (Reinforcement Learning) module</a></li>
-</ul>
-
-<h3>Related Modules</h3>
-
-<p>AI Sorting is built on the powerful <a href="https://www.drupal.org/project/rl"><strong>Reinforcement Learning (RL)</strong></a> module, which provides the core Thompson Sampling algorithm and experiment tracking capabilities. For developers building custom AI-powered features, the RL module offers a comprehensive API for multi-armed bandit experiments.</p>
-
-<p><strong>Ready to supercharge your content marketing?</strong> Install AI Sorting and turn your static content lists into intelligent, self-optimizing marketing tools that work 24/7 to maximize engagement.</p>
